Added LICENSE.txt file
The ReadMe.txt file from Crunchyroll Downloader Toolkit DX specifies
this license for at least ultimate.py and decode.py files. I don't know
which licenses applies for the rest of the files on this repository
(since according to the ReadMe.txt, these files are found on Web), but
I don't think the we should care too much.